Rename CTA components
